Preparations for enabling builds on arm64.

+++ .github/build/macos/madx.sh
@@ -5,24 +5,26 @@ set -ex
 # Must be run from the root the directory of the MAD-X sources.
 # Builds in './build' and installs to './dist'.
 
-rm -rf build
 mkdir build
 cd build
 
-pip3 install --upgrade cmake
-cmake .. \
-    -DCMAKE_POLICY_DEFAULT_CMP0077=NEW \
-    -DCMAKE_POLICY_DEFAULT_CMP0042=NEW \
-    -DCMAKE_OSX_ARCHITECTURES=x86_64 \
-    -DCMAKE_C_COMPILER=gcc-9 \
-    -DCMAKE_CXX_COMPILER=g++-9 \
-    -DCMAKE_Fortran_COMPILER=gfortran-9 \
-    -DBUILD_SHARED_LIBS=OFF \
-    -DMADX_STATIC=OFF \
-    -DCMAKE_INSTALL_PREFIX=../dist \
-    -DCMAKE_BUILD_TYPE=Release \
-    -DMADX_INSTALL_DOC=OFF \
-    -DMADX_ONLINE=OFF \
-    -DMADX_FORCE_32=OFF \
-    -DMADX_X11=OFF
-cmake --build . --target install
+if [[ ! -f CMakeCache.txt ]]; then
+    pip3 install --upgrade cmake
+    cmake .. \
+        -DCMAKE_POLICY_DEFAULT_CMP0077=NEW \
+        -DCMAKE_POLICY_DEFAULT_CMP0042=NEW \
+        -DCMAKE_OSX_ARCHITECTURES=x86_64 \
+        -DCMAKE_C_COMPILER=gcc-9 \
+        -DCMAKE_CXX_COMPILER=g++-9 \
+        -DCMAKE_Fortran_COMPILER=gfortran-9 \
+        -DBUILD_SHARED_LIBS=OFF \
+        -DMADX_STATIC=OFF \
+        -DCMAKE_INSTALL_PREFIX=../dist \
+        -DCMAKE_BUILD_TYPE=Release \
+        -DMADX_INSTALL_DOC=OFF \
+        -DMADX_ONLINE=OFF \
+        -DMADX_FORCE_32=OFF \
+        -DMADX_X11=OFF
+fi
+
+cmake --build . --target install -j
